backend/requirements.txt
asgiref==3.3.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
